matcher-server/requirements_matching.txt
pillow
imagehash
matplotlib
